Decode utf-8 before presentation
This is needed when the strings are encoded in utf8. The encoding is needed for special characters (as simple as é) and smileys. 
In my use case, I only need title, body and big_text (not familiar with the new additions: action, etc.) but any string in the notification could need decoding. 
String X = bundle.getString("X")
X= URLDecoder.decode( X, "UTF-8" );

Medium term, it could be nice to have a tag for encoding in the initial message that would trigger or not a decoding.